+++ R/sf-path.R
@@ -1,6 +1,7 @@
 
 #' @importFrom purrr map_df
-sc_atom <- function(x, ...) faster_as_tibble(list(ncoords_= nrow(x), path_ = sc_uid()))
+sc_atom <- function(x, ...) faster_as_tibble(list(ncoords_= nrow(x), 
+                                                  path_ = sc_uid(1L)))
 sc_list <- function(x) {
   dplyr::bind_rows(lapply(x, sc_atom))
 }